requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
google-generativeai
python-dotenv==1.0.1
pytest